Desktop/ai_career_aid/requirements.txt
Flask
requests
python-dotenv
PyMuPDF
authlib